# wac/backend/src/main/java/dev/pioruocco/wacchat/moderation/BlockedUserResponse.java
package dev.pioruocco.wacchat.moderation;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BlockedUserResponse {

    private String id;
    private String username;
    private String avatarUrl;
}