# wac/frontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/api/v1/users/username":{"put":{"tags":["User"],"operationId":"updateUsername","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/users/{id}/report":{"post":{"tags":["Moderation"],"operationId":"reportUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReportRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/users/{id}/block":{"post":{"tags":["Moderation"],"operationId":"blockUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Moderation"],"operationId":"unblockUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/me/avatar":{"post":{"tags":["User"],"operationId":"uploadAvatar","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["User"],"operationId":"deleteAvatar","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/messages":{"post":{"tags":["Message"],"operationId":"saveMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}},"required":true},"responses":{"201":{"description":"Created"}}},"patch":{"tags":["Message"],"operationId":"setMessageToSeen","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted"}}}},"/api/v1/messages/upload-media":{"post":{"tags":["Message"],"operationId":"uploadMedia","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created"}}}},"/api/v1/internal/sessions/validate":{"post":{"tags":["session-validation-controller"],"operationId":"validate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionValidationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionValidationResponse"}}}}}}},"/api/v1/chats":{"get":{"tags":["Chat"],"operationId":"getChatsByReceiver","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"post":{"tags":["Chat"],"operationId":"createChat","parameters":[{"name":"receiver-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StringResponse"}}}}}}},"/api/v1/chats/{chatId}/reject":{"patch":{"tags":["Chat"],"operationId":"rejectChat","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/chats/{chatId}/favorite":{"patch":{"tags":["Chat"],"operationId":"toggleFavorite","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/v1/chats/{chatId}/accept":{"patch":{"tags":["Chat"],"operationId":"acceptChat","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["User"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["User"],"operationId":"getMe","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/users/check-username":{"get":{"tags":["User"],"operationId":"checkUsername","parameters":[{"name":"value","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/v1/users/blocked":{"get":{"tags":["Moderation"],"operationId":"getBlockedUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlockedUserResponse"}}}}}}}},"/api/v1/messages/chat/{chat-id}":{"get":{"tags":["Message"],"operationId":"getAllMessages","parameters":[{"name":"chat-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}},"/api/v1/users/me/session":{"delete":{"tags":["User"],"operationId":"clearSession","parameters":[{"name":"X-Tab-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserRequest":{"required":["username"],"type":"object","properties":{"username":{"maxLength":20,"minLength":3,"pattern":"^[a-z0-9_.-]+$","type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"},"lastSeen":{"type":"string","format":"date-time"},"avatarUrl":{"type":"string"},"online":{"type":"boolean"}}},"UserReportRequest":{"required":["reason"],"type":"object","properties":{"reason":{"type":"string","enum":["SPAM","HARASSMENT","INAPPROPRIATE_CONTENT","FAKE_PROFILE","OTHER"]},"details":{"maxLength":1000,"minLength":0,"type":"string"}}},"MessageRequest":{"required":["chatId","content","type"],"type":"object","properties":{"content":{"maxLength":4096,"minLength":0,"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","VIDEO","AUDIO"]},"chatId":{"type":"string"}}},"SessionValidationRequest":{"type":"object","properties":{"userId":{"type":"string"},"tabId":{"type":"string"}}},"SessionValidationResponse":{"type":"object","properties":{"conflicting":{"type":"boolean"}}},"StringResponse":{"type":"object","properties":{"response":{"type":"string"}}},"BlockedUserResponse":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"avatarUrl":{"type":"string"}}},"MessageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","VIDEO","AUDIO"]},"state":{"type":"string","enum":["SENT","SEEN"]},"senderId":{"type":"string"},"receiverId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"media":{"type":"array","items":{"type":"string"}}}},"ChatResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"unreadCount":{"type":"integer","format":"int64"},"lastMessage":{"type":"string"},"lastMessageType":{"type":"string","enum":["TEXT","IMAGE","VIDEO","AUDIO"]},"lastMessageTime":{"type":"string","format":"date-time"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"avatarUrl":{"type":"string"},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED"]},"pendingMessageCount":{"type":"integer","format":"int32"},"favorite":{"type":"boolean"},"recipientOnline":{"type":"boolean"}}}},"securitySchemes":{"keycloak":{"type":"oauth2","in":"header","scheme":"bearer","bearerFormat":"JWT","flows":{"password":{"authorizationUrl":"http://localhost:8180/realms/wacchat/protocol/openid-connect/auth","tokenUrl":"http://localhost:8180/realms/wacchat/protocol/openid-connect/token","scopes":{}}}}}}}